# Make Head.symbols map with size hint