Make `Head.symbols` map with size hint
to reduce the number of times the map is resized

---
Benchcmp for [BenchmarkCompaction](https://github.com/prometheus/tsdb/blob/master/compact_test.go#L775)
```
benchmark                                                                                   old ns/op      new ns/op      delta
BenchmarkCompaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=101-4        5840453439     5412773863     -7.32%
BenchmarkCompaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=1001-4       8872856728     8539795503     -3.75%
BenchmarkCompaction/type=vertical,blocks=4,series=10000,samplesPerSeriesPerBlock=101-4      5820110035     5582715232     -4.08%
BenchmarkCompaction/type=vertical,blocks=4,series=10000,samplesPerSeriesPerBlock=1001-4     8704204708     7755268645     -10.90%
```

// Skip `case[2:4]` `case[6:8]`, these cases take a long time to run on my Mac (: